Clean and fixup dllimports.py
